% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_required_arg}
\alias{check_required_arg}
\title{Simple Check}
\usage{
check_required_arg(arg, stop_message)
}
\arguments{
\item{arg}{actual variable\cr{}
give the variable to be used and checked}

\item{stop_message}{{\code{character(1)}}\cr{}
The \code{stop()} message to display explaining why this
argument is needed}
}
\description{
Check that argument is not null nor equal to ""
}
