% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_biogeo_status.R
\name{rt_biogeo_status}
\alias{rt_biogeo_status}
\title{Retrieve biogeographic statuses used in TAXREF}
\usage{
rt_biogeo_status(biogeo_id = NULL)
}
\arguments{
\item{biogeo_id}{{\code{NULL} or \code{character(1)} [default = \code{NULL}]}\cr{}
A biogeographic status id used in TAXREF}
}
\description{
If the function is used without arguments returns the list of biogeographic
statuses used in TAXREF
}
\examples{
\dontrun{rt_biogeo_status(biogeo_id = "P")}

}
