% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_biogeo_status.R
\name{rt_biogeo_status_id}
\alias{rt_biogeo_status_id}
\title{Retrieve a biogeographic status used in TaxRef}
\usage{
rt_biogeo_status_id(biogeo_id)
}
\arguments{
\item{biogeo_id}{{\code{character(1)}}\cr{}
A biogeographic status used in TaxRef see
\code{\link[=rt_biogeo_status]{rt_biogeo_status()}}}
}
\description{
Retrieve a biogeographic status used in TaxRef
}
\examples{
rt_biogeo_status_id("X")

}
