% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_external_db.R
\name{rt_external_db_id}
\alias{rt_external_db_id}
\title{Retrieve an external database referenced in TaxRef}
\usage{
rt_external_db_id(db_id)
}
\arguments{
\item{db_id}{{\code{character(1)}}\cr{}
The id of an external db referenced in TaxRef
\code{\link[=rt_biogeo_status]{rt_biogeo_status()}}}
}
\description{
Retrieve an external database referenced in TaxRef
}
