% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rt_flatten_query}
\alias{rt_flatten_query}
\title{Flatten query}
\usage{
rt_flatten_query(query)
}
\arguments{
\item{query}{{\code{list(1)}}\cr{}
a list that represents the query argument from \code{rt_GET()}}
}
\description{
Flatten a query, meaning that when an argument of the query is an array
the function gives back a list with as many elements in the array named the
same.
}
