% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_habitats.R
\name{rt_habitats_id}
\alias{rt_habitats_id}
\title{Retrieve a habitat used in TaxRef}
\usage{
rt_habitats_id(habitat_id)
}
\arguments{
\item{habitat_id}{{\code{character(1)}}\cr{}
The id of a habitat referenced in TaxRef
\code{\link[=rt_habitats]{rt_habitats()}} see section below}
}
\description{
Retrieve a habitat used in TaxRef
}
\section{Habitats}{

\itemize{
\item \code{1} = Marine
\item \code{2} = Freshwater
\item \code{3} = Terrestrial
\item \code{4} = Marine and freshwater
\item \code{5} = Marine and Terrestrial
\item \code{6} = Brackish water
\item \code{7} = Continental (terrestrial and/or freshwater)
\item \code{8} = Continental (terrestrial and freshwater)
}
}

\examples{
rt_habitats_id(habitat_id = 8)

}
