% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_interactions_search.R
\name{rt_interactions_search}
\alias{rt_interactions_search}
\title{Search Interactions based on a list of criteria}
\usage{
rt_interactions_search(id = NULL, interaction_id = NULL,
  opgroup_id = NULL, page = 1, size = 2000)
}
\arguments{
\item{id}{{\code{integer(1)}}\cr{}
an integer giving the id of the taxon in TaxRef (also called
\code{cdNom})}

\item{interaction_id}{{\code{character(1)}}\cr{}
A character identifying the type of interactions (see
\code{\link[=rt_interactions_type]{rt_interactions_type()}} for a list of interactions types)}

\item{opgroup_id}{{\code{numeric(1)}}\cr{}
the id of an operational group referenced in TaxRef see
\code{\link[=rt_operational_groups]{rt_operational_groups()}}}

\item{page}{{\code{integer(1)}; default = \code{1}}\cr{}
The page number returned}

\item{size}{{\code{integer(1)}; default = \code{50000}}\cr{}
The number of lines returned per page (maximum = 50000)}
}
\description{
Search Interactions based on a list of criteria
}
