% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_interactions_type.R
\name{rt_interactions_type}
\alias{rt_interactions_type}
\title{Retrieve all Interactions Types}
\usage{
rt_interactions_type()
}
\description{
Retrieve all Interactions Types
}
\seealso{
\code{\link[=rt_interactions_find_type]{rt_interactions_find_type()}} to retrieve all information by
interactions
}
