% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_interactions_type.R
\name{rt_interactions_type_id}
\alias{rt_interactions_type_id}
\title{Retrieve a interactions Types based on its id}
\usage{
rt_interactions_type_id(interaction_id)
}
\arguments{
\item{interaction_id}{{\code{character(1)}}\cr{}
A character identifying the type of interactions (see
\code{\link[=rt_interactions_type]{rt_interactions_type()}} for a list of interactions types)}
}
\description{
Retrieve a interactions Types based on its id
}
\examples{
rt_interactions_type_id("PREDATEUR")
}
\seealso{
see \code{\link[=rt_interactions_type]{rt_interactions_type()}} for a list of interactions types
}
