% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_langages.R
\name{rt_langages_id}
\alias{rt_langages_id}
\title{Retrieve a habitat used in TaxRef}
\usage{
rt_langages_id(langages_id)
}
\arguments{
\item{langages_id}{{\code{character(1)}}\cr{}
three-letter lowercase code that is the id of a langage
referenced in TaxRef see \code{\link[=rt_langages]{rt_langages()}}}
}
\description{
Retrieve a habitat used in TaxRef
}
