% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_languages.R
\name{rt_languages}
\alias{rt_languages}
\title{Retrieve the languages used in TAXREF}
\usage{
rt_languages(languages_id = NULL)
}
\arguments{
\item{languages_id}{{\code{character(1)}, default = \code{NULL}}\cr{}
three-letter lowercase code that is the id of a language
referenced in TAXREF}
}
\description{
If used without argument returns the list of languages used in TAXREF
otherwise the specific
}
\examples{
# Retrieve the TAXREF definition in the Wayana Language
rt_languages(languages_id = "way")

}
