% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_operational_groups.R
\name{rt_operational_groups}
\alias{rt_operational_groups}
\title{Retrieve the list of operational groups used in TaxRef}
\usage{
rt_operational_groups()
}
\description{
Operational Groups are names of groups either taxonomic or vernacular
used across TaxRef in French
}
