% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_operational_groups.R
\name{rt_operational_groups_id}
\alias{rt_operational_groups_id}
\title{Retrieve an operational group used in TaxRef}
\usage{
rt_operational_groups_id(opgroup_id)
}
\arguments{
\item{opgroup_id}{{\code{numeric(1)}}\cr{}
the id of an operational group referenced in TaxRef see
\code{\link[=rt_operational_groups]{rt_operational_groups()}}}
}
\description{
Retrieve an operational group used in TaxRef
}
