% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_source_uses.R
\name{rt_source_uses_id}
\alias{rt_source_uses_id}
\title{Retrieve a source use in TaxRef}
\usage{
rt_source_uses_id(usage_id)
}
\arguments{
\item{usage_id}{{\code{numeric(1)}}\cr{}
the id of a source use in TaxRef see
\code{\link[=rt_source_uses]{rt_source_uses()}}}
}
\description{
Retrieve a source use in TaxRef
}
