% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_sources_find_term.R
\name{rt_sources_find_term}
\alias{rt_sources_find_term}
\title{Search a source from a term}
\usage{
rt_sources_find_term(term = NULL)
}
\arguments{
\item{term}{{\code{character(1)}}\cr{}
A term contained in the source}
}
\description{
Search a source from a term
}
\examples{
\dontrun{
rt_sources_find_term(term = "Ophrys")
}
}
