% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_sources_match.R
\name{rt_sources_match}
\alias{rt_sources_match}
\title{Search a source from a citation by fuzzy matchin}
\usage{
rt_sources_match(citation = NULL)
}
\arguments{
\item{citation}{{\code{character(1)}}\cr{}
A (partial) citation}
}
\description{
Search a source from a citation by fuzzy matchin
}
