% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_status_opgroup.R
\name{rt_status_opgroup}
\alias{rt_status_opgroup}
\title{Retrieve the status of all Taxa in an Operational Group}
\usage{
rt_status_opgroup(id, page = 1, size = 2000)
}
\arguments{
\item{page}{{\code{integer(1)}; default = \code{1}}\cr{}
The page number returned}

\item{size}{{\code{integer(1)}; default = \code{50000}}\cr{}
The number of lines returned per page (maximum = 50000)}

\item{opgroup_id}{{\code{integer(1)}}\cr{}
identifier of the operational group (see the list of
operational groups \code{\link[=rt_operational_groups]{rt_operational_groups()}})}
}
\description{
Retrieve the status of all Taxa in an Operational Group
}
\seealso{
the list of operational groups \code{\link[=rt_operational_groups]{rt_operational_groups()}}
}
