% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_status_search.R
\name{rt_status_search}
\alias{rt_status_search}
\title{Search statuses based on a list of criteria}
\usage{
rt_status_search(id = NULL, opgroup_id = NULL, location_id = NULL,
  latitude = NULL, longitude = NULL, page = 1, size = 2000,
  detail = FALSE)
}
\arguments{
\item{id}{{\code{integer(1)}}\cr{}
an integer giving the id of the taxon in TaxRef (also called
\code{cdNom})}

\item{opgroup_id}{{\code{numeric(1)}}\cr{}
the id of an operational group referenced in TaxRef see
\code{\link[=rt_operational_groups]{rt_operational_groups()}}}

\item{location_id}{{\code{character(1)}}\cr{}
a string identifying a location}

\item{latitude}{{\code{numeric(1)}}\cr{}
a latitude value}

\item{longitude}{{\code{numeric(1)}}\cr{}
a longitude value}

\item{page}{{\code{integer(1)}; default = \code{1}}\cr{}
The page number returned}

\item{size}{{\code{integer(1)}; default = \code{50000}}\cr{}
The number of lines returned per page (maximum = 50000)}

\item{detail}{{\code{logical(1)} default = \code{FALSE}}\cr{}
Should the table be a long table detailing all the statuses or
a compact table?}
}
\description{
Search statuses based on a list of criteria
}
