% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_status_type.R
\name{rt_status_type}
\alias{rt_status_type}
\title{Retrieve all Status Types}
\usage{
rt_status_type()
}
\description{
Retrieve all Status Types
}
\seealso{
\code{\link[=rt_status_find_type]{rt_status_find_type()}} to retrieve all information by status
}
