% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_status_type.R
\name{rt_status_type_id}
\alias{rt_status_type_id}
\title{Retrieve a Status Types based on its id}
\usage{
rt_status_type_id(status_id)
}
\arguments{
\item{status_id}{{\code{character(1)}}\cr{}
A character identifying the type of status (see
\code{\link[=rt_status_type]{rt_status_type()}} for a list of status types)}
}
\description{
Retrieve a Status Types based on its id
}
\examples{
rt_status_type_id("DH")
}
\seealso{
see \code{\link[=rt_status_type]{rt_status_type()}} for a list of status types
}
