% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_autocomplete.R
\name{rt_taxa_autocomplete}
\alias{rt_taxa_autocomplete}
\title{Retrieve Taxon ID from partially matching its Name}
\usage{
rt_taxa_autocomplete(
  term = NULL,
  territories = NULL,
  rank = NULL,
  domain = NULL,
  page = 1,
  size = 200
)
}
\arguments{
\item{term}{{\code{character(1)} [\strong{required}]}\cr{}
the beginning of the scientific name of the searched taxa}

\item{territories}{{\code{NULL} or \verb{character(1+)} [default = \code{NULL}]}\cr{}
One or more territories where species is present:
\itemize{
\item \code{"fr"} = Mainland France
\item \code{"gf"} = French Guiana
\item \code{"gua"} = Guadeloupe
\item \code{"mar"} = Martinique
\item \code{"sm"} = Saint Martin
\item \code{"sb"} = Saint Barthélemy
\item \code{"spm"} = Saint Pierre and Miquelon
\item \code{"epa"} = Scattered Islands in the Indian Ocean
\item \code{"may"} = Mayotte
\item \code{"reu"} = Réunion
\item \code{"sa"} = French Southern and Antarctic Lands
\item \code{"ta"} = Adélie Land
\item \code{"nc"} = New Caledonia
\item \code{"wf"} = Wallis and Futuna
\item \code{"pf"} = French Polynesia
\item \code{"cli"} = Clipperton Island
}}

\item{rank}{{\code{NULL} or \verb{character(1+)} [default = \code{NULL}]}\cr{}
One of more rank of the searched taxa (see \code{\link[=rt_taxo_ranks]{rt_taxo_ranks()}} for the list
of all ranks available):
\itemize{
\item \code{Dumm}= Domain
\item \code{KD}= Kingdom
\item \code{PH}= Phylum
\item \code{CL}= Class
\item \code{OR}= Order
\item \code{FM}= Family
\item \code{SBFM}= Subfamily
\item \code{TR}= Tribe
\item \code{GN}= Genus
\item \code{AGES}= Aggregate
\item \code{ES}= Species
\item \code{SSES}= Subspecies
\item \code{NAT}= Natio
\item \code{VAR}= Variety
\item \code{SVAR}= Sub-variety
\item \code{FO}= Form
\item \code{SSFO}= Sub-form
\item \code{RACE}= Race
\item \code{CAR}= Cultivar
\item \code{AB}= Abberatio
}}

\item{domain}{{\code{NULL} or \code{character(1)} [default = \code{NULL}]}\cr{}
The domain where the species is found:
\itemize{
\item \code{"marin"} = Marine
\item \code{"continental"} = Continental
}}

\item{page}{{\code{integer(1)} [default = \code{1}]}\cr{}
The page number returned}

\item{size}{{\code{integer(1)} [default = \code{200}]}\cr{}
The number of lines returned per page (maximum = 200)}
}
\description{
Retrieve Taxon ID from partially matching its Name
}
\examples{
\dontrun{
rt_taxa_autocomplete(term = "Bradypus")
}

}
\seealso{
the list of taxonomic ranks avaible through \code{\link[=rt_taxo_ranks]{rt_taxo_ranks()}}
}
