% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_children.R
\name{rt_taxa_children}
\alias{rt_taxa_children}
\title{Retrieve a taxon children}
\usage{
rt_taxa_children(id)
}
\arguments{
\item{id}{{\code{integer(1)}}\cr{}
an integer giving the id of the taxon in TAXREF (also called
\code{cdNom})}
}
\description{
Retrieve a taxon children
}
\examples{
\dontrun{rt_taxa_children(id = 442585)}

}
