% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_externalids.R
\name{rt_taxa_externalids}
\alias{rt_taxa_externalids}
\title{Retrieve IDs of the taxon in external databases}
\usage{
rt_taxa_externalids(id)
}
\arguments{
\item{id}{{\code{integer(1)}}\cr{}
an integer giving the id of the taxon in TAXREF (also called
\code{cdNom})}
}
\description{
Retrieve IDs of the taxon in external databases
}
\examples{
\dontrun{rt_taxa_externalids(id = 443800)}

}
