% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_from_externalids.R
\name{rt_taxa_from_externalids}
\alias{rt_taxa_from_externalids}
\title{Retrieve Taxon using the ids from external databases}
\usage{
rt_taxa_from_externalids(external_db = NULL, external_id = NULL)
}
\arguments{
\item{external_db}{{\code{character(1)}}\cr{}
The name of an external database see \code{\link[=rt_external_db]{rt_external_db()}}
for the list of external database names}

\item{external_id}{{\code{character(1)}}\cr{}
The identifier of the taxon in the specified external
database}
}
\description{
Retrieve Taxon using the ids from external databases
}
\examples{
rt_taxa_from_externalids("TROPICOS", 26600197)
}
\seealso{
\code{\link[=rt_external_db]{rt_external_db()}} for the list of external databases linked to
TaxRef; \code{\link[=rt_taxa_externalids]{rt_taxa_externalids()}} to get the external ids associated with
a taxon in TaxRef
}
