% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_fuzzymatch.R
\name{rt_taxa_fuzzymatch}
\alias{rt_taxa_fuzzymatch}
\title{Find taxa by fuzzy matching their scientific Name}
\usage{
rt_taxa_fuzzymatch(sciname = NULL)
}
\arguments{
\item{sciname}{{\code{character(1)}, \strong{required}}\cr{}
a string giving the scientific name of the taxon with or
without the author name}
}
\description{
Find taxa by fuzzy matching their scientific Name
}
