% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_interactions.R
\name{rt_taxa_interactions}
\alias{rt_taxa_interactions}
\title{Retrieve Interactions of taxon with other taxa}
\usage{
rt_taxa_interactions(id = NULL)
}
\arguments{
\item{id}{{\code{integer(1)} [\strong{required}]}\cr{}
an integer giving the id of the taxon in TAXREF (also called
\code{cdNom})}
}
\description{
Retrieve Interactions of taxon with other taxa
}
\examples{
\dontrun{rt_taxa_interactions(id = 110335)}

}
