% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_media.R
\name{rt_taxa_media}
\alias{rt_taxa_media}
\title{Retrieve Media (photos & others) Associated with a Taxon}
\usage{
rt_taxa_media(id)
}
\arguments{
\item{id}{{\code{integer(1)}}\cr{}
an integer giving the id of the taxon in TaxRef (also called
\code{cdNom})}
}
\description{
Retrieve Media (photos & others) Associated with a Taxon
}
