% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_parent.R
\name{rt_taxa_parent}
\alias{rt_taxa_parent}
\title{Retrieve classification above the taxon}
\usage{
rt_taxa_parent(id)
}
\arguments{
\item{id}{{\code{integer(1)}}\cr{}
an integer giving the id of the taxon in TAXREF (also called
\code{cdNom})}
}
\description{
Retrieve classification above the taxon
}
\examples{
rt_taxa_parent(id = 443800)

}
