% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_search.R
\name{rt_taxa_search}
\alias{rt_taxa_search}
\title{Search for a taxon information}
\usage{
rt_taxa_search(
  id = NULL,
  sciname = NULL,
  fr_name = NULL,
  en_name = NULL,
  rank = NULL,
  territories = NULL,
  domain = NULL,
  habitats = NULL,
  vernacular = NULL,
  version = NULL,
  page = 1,
  size = 5000
)
}
\arguments{
\item{id}{{\code{NULL} or \verb{integer(1+)}}\cr{}
One or more integer giving the id(s) of taxa in TAXREF
(also called \code{cdNom})}

\item{sciname}{{\code{NULL} or \verb{character(1+)}}\cr{}
One or more scientific name of searched taxa}

\item{fr_name}{{\code{NULL} or \verb{character(1+)}}\cr{}
One or more common name(s) in French of searched taxa}

\item{en_name}{{\code{NULL} or \verb{character(1+)}}\cr{}
One or more common name(s) in English of searched taxa}

\item{rank}{{\code{NULL} or \verb{character(1+)}}\cr{}
One of more rank of the searched taxa (see \code{\link[=rt_taxo_ranks]{rt_taxo_ranks()}} for the list
of all ranks available):
\itemize{
\item \code{Dumm}= Domain
\item \code{KD}= Kingdom
\item \code{PH}= Phylum
\item \code{CL}= Class
\item \code{OR}= Order
\item \code{FM}= Family
\item \code{SBFM}= Subfamily
\item \code{TR}= Tribe
\item \code{GN}= Genus
\item \code{AGES}= Aggregate
\item \code{ES}= Species
\item \code{SSES}= Subspecies
\item \code{NAT}= Natio
\item \code{VAR}= Variety
\item \code{SVAR}= Sub-variety
\item \code{FO}= Form
\item \code{SSFO}= Sub-form
\item \code{RACE}= Race
\item \code{CAR}= Cultivar
\item \code{AB}= Abberatio
}}

\item{territories}{{\code{NULL} or \verb{character(1+)}}\cr{}
One or more territories where species is present:
\itemize{
\item \code{"fr"} = Mainland France
\item \code{"gf"} = French Guiana
\item \code{"gua"} = Guadeloupe
\item \code{"mar"} = Martinique
\item \code{"sm"} = Saint Martin
\item \code{"sb"} = Saint Barthélemy
\item \code{"spm"} = Saint Pierre and Miquelon
\item \code{"epa"} = Scattered Islands in the Indian Ocean
\item \code{"may"} = Mayotte
\item \code{"reu"} = Réunion
\item \code{"sa"} = French Southern and Antarctic Lands
\item \code{"ta"} = Adélie Land
\item \code{"nc"} = New Caledonia
\item \code{"wf"} = Wallis and Futuna
\item \code{"pf"} = French Polynesia
\item \code{"cli"} = Clipperton Island
}}

\item{domain}{{\code{NULL} or \code{character(1)}}\cr{}
The domain where the species is found:
\itemize{
\item \code{"marin"} = Marine
\item \code{"continental"} = Continental
}}

\item{habitats}{{\code{NULL} or \verb{integer(1+)}}\cr{}
One or more habitats where the species is found (see \code{\link[=rt_habitats]{rt_habitats()}} for a
list of all habitats):
\itemize{
\item \code{1} = Marine
\item \code{2} = Freshwater
\item \code{3} = Terrestrial
\item \code{4} = Marine and freshwater
\item \code{5} = Marine and Terrestrial
\item \code{6} = Brackish water
\item \code{7} = Continental (terrestrial and/or freshwater)
\item \code{8} = Continental (terrestrial and freshwater)
}}

\item{vernacular}{{\code{NULL} or \verb{character(1+)}}\cr{}
Name of a vernacular group
(see the list using \code{\link[=rt_vernacular_groups]{rt_vernacular_groups()}})}

\item{version}{{\code{NULL} or \code{character(1)}}\cr{}
A string indicating which version of TAXREF should be queried.
Should be of the form \code{"2.0"} up to \code{"15.0"}. You can check
the available TAXREF versions with \code{\link[=rt_taxref_versions]{rt_taxref_versions()}}.}

\item{page}{{\code{integer(1)} [default = \code{1}]}\cr{}
The page number returned}

\item{size}{{\code{integer(1)} [default = \code{5000}]}\cr{}
The number of lines returned per page (maximum = 5000)}
}
\description{
Search for a taxon information
}
\examples{
\dontrun{rt_taxa_search(sciname = "Bradypus")}

}
