% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_sources.R
\name{rt_taxa_sources}
\alias{rt_taxa_sources}
\title{Retrieve Sources Associated with a Taxon}
\usage{
rt_taxa_sources(id)
}
\arguments{
\item{id}{{\code{integer(1)}}\cr{}
an integer giving the id of the taxon in TaxRef (also called
\code{cdNom})}
}
\description{
Retrieve Sources Associated with a Taxon
}
\seealso{
\code{\link[=rt_source_uses]{rt_source_uses()}} for types of source use
}
