% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_status.R
\name{rt_taxa_status}
\alias{rt_taxa_status}
\title{Retrieve Taxon Conservation Status}
\usage{
rt_taxa_status(id, detail = FALSE)
}
\arguments{
\item{id}{{\code{integer(1)}}\cr{}
an integer giving the id of the taxon in TaxRef (also called
\code{cdNom})}

\item{detail}{{\code{logical(1)} default = \code{FALSE}}\cr{}
Should the table be a long table detailing all the statuses or
a compact table?}
}
\description{
Retrieve Taxon Conservation Status
}
\examples{
rt_taxa_status(id = 443800, detail = FALSE)
rt_taxa_status(id = 443800, detail = TRUE)

}
