% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_synonyms.R
\name{rt_taxa_synonyms}
\alias{rt_taxa_synonyms}
\title{Retrieve Synonyms associated with a Taxon}
\usage{
rt_taxa_synonyms(id)
}
\arguments{
\item{id}{{\code{integer(1)}}\cr{}
an integer giving the id of the taxon in TAXREF (also called
\code{cdNom})}
}
\description{
Retrieve Synonyms associated with a Taxon
}
\examples{
\dontrun{rt_taxa_synonyms(id = 107003)}

}
