% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_taxrefhistory.R
\name{rt_taxa_taxrefhistory}
\alias{rt_taxa_taxrefhistory}
\title{Retrieve history of a taxon through TAXREF versions}
\usage{
rt_taxa_taxrefhistory(id = NULL)
}
\arguments{
\item{id}{{\code{integer(1)}, \strong{required}}\cr{}
an integer giving the id of the taxon in TAXREF (also called
\code{cdNom})}
}
\description{
Retrieve history of a taxon through TAXREF versions
}
\examples{
\dontrun{rt_taxa_taxrefhistory(id = 443800)}

}
