% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_vernacular.R
\name{rt_taxa_vernacular}
\alias{rt_taxa_vernacular}
\title{Retrieve VernacularNames associated with a Taxon}
\usage{
rt_taxa_vernacular(id)
}
\arguments{
\item{id}{{\code{integer(1)}}\cr{}
an integer giving the id of the taxon in TaxRef (also called
\code{cdNom})}
}
\description{
Retrieve VernacularNames associated with a Taxon
}
