% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxo_ranks.R
\name{rt_taxo_ranks_id}
\alias{rt_taxo_ranks_id}
\title{Retrieve a taxonomic rank used in TaxRef}
\usage{
rt_taxo_ranks_id(taxo_id)
}
\arguments{
\item{taxo_id}{{\code{character(1)}}\cr{}
the id of a taxonomic rank in TaxRef see
\code{\link[=rt_taxo_ranks]{rt_taxo_ranks()}}}
}
\description{
Retrieve a taxonomic rank used in TaxRef
}
\examples{
rt_taxo_ranks_id(taxo_id = "VAR")

}
