% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxref_versions.R
\name{rt_taxref_versions_id}
\alias{rt_taxref_versions_id}
\title{Retrieve information about a TaxRef version}
\usage{
rt_taxref_versions_id(version_id)
}
\arguments{
\item{version_id}{{\code{integer(1)}}\cr{}
the id of a TaxRef version
\code{\link[=rt_taxref_versions]{rt_taxref_versions()}}}
}
\description{
Retrieve information about a TaxRef version
}
