% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_vernacular_groups.R
\name{rt_vernacular_groups}
\alias{rt_vernacular_groups}
\title{Retrieve the list of vernacular groups in TaxRef}
\usage{
rt_vernacular_groups()
}
\description{
Retrieve the list of vernacular groups in TaxRef
}
