% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_biogeo_status.R
\name{rt_biogeo_status}
\alias{rt_biogeo_status}
\title{Retrieve the list of biogeographic statuses used in TaxRef}
\usage{
rt_biogeo_status()
}
\description{
Retrieve the list of biogeographic statuses used in TaxRef
}
