% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_external_db.R
\name{rt_external_db}
\alias{rt_external_db}
\title{Retrieve the list of external DBs referenced in TaxRef}
\usage{
rt_external_db()
}
\description{
Retrieve the list of external DBs referenced in TaxRef
}
