% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_habitats.R
\name{rt_habitats}
\alias{rt_habitats}
\title{Retrieve habitats used in TaxRef}
\usage{
rt_habitats(habitat_id = NULL)
}
\arguments{
\item{habitat_id}{{\code{character(1)}, default = \code{NULL}}\cr{}
The id of a habitat referenced in TaxRef}
}
\description{
If the function is used without arguments returns the list of habitats used
in TaxRef.
}
\section{Habitats}{

\itemize{
\item \code{1} = Marine
\item \code{2} = Freshwater
\item \code{3} = Terrestrial
\item \code{4} = Marine and freshwater
\item \code{5} = Marine and Terrestrial
\item \code{6} = Brackish water
\item \code{7} = Continental (terrestrial and/or freshwater)
\item \code{8} = Continental (terrestrial and freshwater)
}
}

\examples{
rt_habitats(habitat_id = 8)

}
