% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_interactions_search.R
\name{rt_interactions_search}
\alias{rt_interactions_search}
\title{Search Interactions based on a list of criteria}
\usage{
rt_interactions_search(
  id = NULL,
  interaction_id = NULL,
  opgroup_id = NULL,
  page = 1,
  size = 2000
)
}
\arguments{
\item{id}{{\verb{integer(1+)}}\cr{}
One or more integer giving the id(s) of taxa in TAXREF
(also called \code{cdNom})}

\item{interaction_id}{{\code{character(1)}, default = \code{NULL}}\cr{}
A character identifying the type of interactions}

\item{opgroup_id}{{\code{integer(1)}, \strong{required}}\cr{}
identifier of the operational group (see the list of
operational groups \code{\link[=rt_operational_groups]{rt_operational_groups()}})}

\item{page}{{\code{integer(1)}; default = \code{1}}\cr{}
The page number returned}

\item{size}{{\code{integer(1)}; default = \code{2000}}\cr{}
The number of lines returned per page (maximum = 2000)}
}
\description{
Search Interactions based on a list of criteria
}
\examples{
\dontrun{
# Interactors with 'Ophrys apifera' (id = 110335)
rt_interactions_search(id = 110335)
}

}
