% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_interactions_search.R
\name{rt_interactions_search}
\alias{rt_interactions_search}
\title{Search Interactions based on a list of criteria}
\usage{
rt_interactions_search(
  id = NULL,
  interaction_id = NULL,
  opgroup_id = NULL,
  page = 1,
  size = 2000
)
}
\arguments{
\item{id}{{\code{NULL} or \verb{integer(1+)}}\cr{}
One or more integer giving the id(s) of taxa in TAXREF
(also called \code{cdNom})}

\item{interaction_id}{{\code{NULL} or \code{character(1)} [default = \code{NULL}]}\cr{}
A character identifying the type of interactions}

\item{opgroup_id}{{\code{NULL} or \code{numeric(1)} [default = \code{NULL}]}\cr{}
the id of an operational group referenced in TAXREF}

\item{page}{{\code{integer(1)} [default = \code{1}]}\cr{}
The page number returned}

\item{size}{{\code{integer(1)} [default = \code{5000}]}\cr{}
The number of lines returned per page (maximum = 5000)}
}
\description{
Search Interactions based on a list of criteria
}
\examples{
\dontrun{
# Interactors with 'Ophrys apifera' (id = 110335)
rt_interactions_search(id = 110335)
}

}
\seealso{
the list of interaction type \code{\link[=rt_interactions_type]{rt_interactions_type()}} and
the list of operational groups \code{\link[=rt_operational_groups]{rt_operational_groups()}}
}
