% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_interactions_type.R
\name{rt_interactions_type}
\alias{rt_interactions_type}
\title{Retrieve Interactions Types used in TAXREF}
\usage{
rt_interactions_type(interaction_id = NULL)
}
\arguments{
\item{interaction_id}{{\code{character(1)}, default = \code{NULL}}\cr{}
A character identifying the type of interactions}
}
\description{
If used without argument returns the list of all interactions used in TAXREF
}
\examples{
rt_interactions_type(interaction_id = "PREDATEUR")

}
