% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_languages.R
\name{rt_languages_id}
\alias{rt_languages_id}
\title{Retrieve a language used in TaxRef}
\usage{
rt_languages_id(languages_id)
}
\arguments{
\item{languages_id}{{\code{character(1)}}\cr{}
three-letter lowercase code that is the id of a language
referenced in TaxRef see \code{\link[=rt_languages]{rt_languages()}}}
}
\description{
Retrieve a language used in TaxRef
}
\examples{
# Retrieve the TaxRef definition in the Wayana Language
rt_languages_id(languages_id = "way")

}
