% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_operational_groups.R
\name{rt_operational_groups}
\alias{rt_operational_groups}
\title{Retrieve operational groups used in TaxRef}
\usage{
rt_operational_groups(opgroup_id = NULL)
}
\arguments{
\item{opgroup_id}{{\code{numeric(1)}, default = \code{NULL}}\cr{}
the id of an operational group referenced in TaxRef}
}
\description{
Operational Groups are names of groups either taxonomic or vernacular
used across TaxRef in French. If used without argument returns the list of
all operational groups used in TaxRef.
}
\examples{
rt_operational_groups(opgroup_id = 143)
}
