% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_source_uses.R
\name{rt_source_uses}
\alias{rt_source_uses}
\title{Retrieve the list of source uses in TAXREF}
\usage{
rt_source_uses(usage_id = NULL)
}
\arguments{
\item{usage_id}{{\code{NULL} or \code{numeric(1)} [default = \code{NULL}]}\cr{}
the id of a source use in TAXREF}
}
\description{
TAXREF references many publications across its documentation. The different
sources are used for different reasons. This function returns the list of
usages for a source when no arguments are provided, otherwise returns the
specific type of use.
}
\examples{
\dontrun{rt_source_uses(usage_id = 276)}

}
