% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_sources_match.R
\name{rt_sources_match}
\alias{rt_sources_match}
\title{Search a source from a citation by fuzzy matchin}
\usage{
rt_sources_match(citation = NULL)
}
\arguments{
\item{citation}{{\code{character(1)}}\cr{}
A (partial) citation}
}
\description{
Search a source from a citation by fuzzy matchin
}
\examples{
\dontrun{
rt_sources_match(
    paste0("Chapman, T. A. 1909. Callophrys avis a new butterfly on ",
           "southern France. The Entomologist's record and journal of ",
           "variation, 21: 130-131.")
)
}
}
