% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_status_search.R
\name{rt_status_search}
\alias{rt_status_search}
\title{Search statuses based on a list of criteria}
\usage{
rt_status_search(
  id = NULL,
  opgroup_id = NULL,
  location_id = NULL,
  latitude = NULL,
  longitude = NULL,
  page = 1,
  size = 2000,
  detail = FALSE
)
}
\arguments{
\item{id}{{\code{integer(1)}, \strong{required}}\cr{}
an integer giving the id of the taxon in TAXREF (also called
\code{cdNom})}

\item{opgroup_id}{{\code{integer(1)}}\cr{}
identifier of the operational group (see the list of
operational groups \code{\link[=rt_operational_groups]{rt_operational_groups()}})}

\item{location_id}{{\code{character(1)}}\cr{}
a string identifying a location}

\item{latitude}{{\code{numeric(1)}}\cr{}
a latitude value}

\item{longitude}{{\code{numeric(1)}}\cr{}
a longitude value}

\item{page}{{\code{integer(1)}; default = \code{1}}\cr{}
The page number returned}

\item{size}{{\code{integer(1)}; default = \code{2000}}\cr{}
The number of lines returned per page (maximum = 2000)}

\item{detail}{{\code{logical(1)} default = \code{FALSE}}\cr{}
Should the table be a long table detailing all the statuses or
a compact table?}
}
\description{
Search statuses based on a list of criteria
}
\examples{
\dontrun{
rt_status_search(id = 443800, detail = FALSE)
rt_status_search(id = 443800, detail = TRUE)
}

}
