% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_status_type.R
\name{rt_status_type}
\alias{rt_status_type}
\title{Retrieve Status Types Used in TAXREF}
\usage{
rt_status_type(status_id = NULL)
}
\arguments{
\item{status_id}{{\code{NULL} or \code{character(1)} [default = \code{NULL}]}\cr{}
A character identifying the type of status}
}
\description{
If the function is provided without arguments returns the entire list of
status types used in TAXREF.
}
