% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_factsheet.R
\name{rt_taxa_factsheet}
\alias{rt_taxa_factsheet}
\title{Get factsheet for taxa}
\usage{
rt_taxa_factsheet(id)
}
\arguments{
\item{id}{{\code{integer(1)}}\cr{}
an integer giving the id of the taxon in TaxRef (also called
\code{cdNom})}
}
\description{
For some taxa, a factsheet has been written. This text, often in French,
describes the general ecology and some diagnosis of the submitted taxon.
}
\details{
Beware, many taxa have no corresponding factsheet.
}
\examples{
rt_taxa_factsheet(4540)
}
