% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_id.R
\name{rt_taxa_id}
\alias{rt_taxa_id}
\title{Retrieve a taxon information using its id}
\usage{
rt_taxa_id(id)
}
\arguments{
\item{id}{{\code{integer(1)}}\cr{}
an integer giving the id of the taxon in TaxRef (also called
\code{cdNom})}
}
\description{
Retrieve a taxon information using its id
}
\examples{
rt_taxa_id(id = 443800)
rt_taxa_id(id = 110335)
}
