% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_media.R
\name{rt_taxa_media}
\alias{rt_taxa_media}
\title{Retrieve Media (photos & others) Associated with a Taxon}
\usage{
rt_taxa_media(id = NULL)
}
\arguments{
\item{id}{{\code{integer(1)} [\strong{required}]}\cr{}
an integer giving the id of the taxon in TAXREF (also called
\code{cdNom})}
}
\description{
Retrieve Media (photos & others) Associated with a Taxon
}
\examples{
\dontrun{rt_taxa_media(id = 443800)}

}
