% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_search.R
\name{rt_taxa_search}
\alias{rt_taxa_search}
\title{Title}
\usage{
rt_taxa_search(sciname = NULL, fr_name = NULL, en_name = NULL,
  rank = NULL, territories = NULL, domain = NULL, habitats = NULL,
  vernacular = NULL)
}
\arguments{
\item{vernacular}{}
}
\description{
Title
}
