% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_taxa_vernacular.R
\name{rt_taxa_vernacular}
\alias{rt_taxa_vernacular}
\title{Retrieve VernacularNames associated with a Taxon}
\usage{
rt_taxa_vernacular(id = NULL)
}
\arguments{
\item{id}{{\code{integer(1)}, \strong{required}}\cr{}
an integer giving the id of the taxon in TAXREF (also called
\code{cdNom})}
}
\description{
Retrieve VernacularNames associated with a Taxon
}
\examples{
\dontrun{rt_taxa_vernacular(id = 443800)}

}
\seealso{
\code{\link[=rt_languages]{rt_languages()}} for a list of languages referenced in TAXREF
}
